fix: product-price-type bug in backend; fix: product-card bug

+++ Backend/src/graphql/types/product.ts
@@ -28,7 +28,7 @@ const ProductTypeDef = `#graphql
         image: String,
         images: [String!],
         brand: String,
-        price: Int!,
+        price: Float!,
         category: Category!,
         countInStock: Int!,
         rating: Int,
